Fix for missing Battery tab in System Settings
Rename the PM Profile in the FACP ACPI table from "8" (Tablet) to "2" (Laptop). This makes the battery tab appear again in System Settings on macOS Ventura and Sonoma.

macOS doesn't know what to do with a "8" (Tablet) PM Profile, so it treats the device like a desktop PC. Renaming the PM Profile to "2" (Laptop) in the FACP ACPI table makes the Battery tab appear again in the System Settings.